+++ plotly/io/sg_scraper.py
@@ -1,0 +1,2 @@
+# ruff: noqa: F401
+from ._sg_scraper import plotly_sg_scraper, reset_renderer